+++ influx/querybuilder.go
@@ -104,7 +104,7 @@ func (q Query) OnSubquery(subquery Query) Query {
 // Field adds the given field to the list of fields with the given aggregation method applied. It
 // is possible to add multiple fields with the same name but is highly discouraged.
 func (q Query) Field(fieldname, aggregationMethod string) Query {
-	q.fields = append(q.fields, fmt.Sprintf("%s(\"%s\") AS %s", aggregationMethod, fieldname, fieldname))
+	q.fields = append(q.fields, fmt.Sprintf("%s(\"%s\") AS \"%s\"", aggregationMethod, fieldname, fieldname))
 	return q
 }
 
@@ -114,7 +114,17 @@ func (q Query) Median(fieldname string, aliases ...string) Query {
 	if len(aliases) > 0 {
 		alias = aliases[0]
 	}
-	q.fields = append(q.fields, fmt.Sprintf("median(\"%s\") AS %s", fieldname, alias))
+	q.fields = append(q.fields, fmt.Sprintf("median(\"%s\") AS \"%s\"", fieldname, alias))
+	return q
+}
+
+// Min adds the field `min` to the query.
+func (q Query) Min(fieldname string, aliases ...string) Query {
+	alias := fieldname
+	if len(aliases) > 0 {
+		alias = aliases[0]
+	}
+	q.fields = append(q.fields, fmt.Sprintf("min(\"%s\") AS \"%s\"", fieldname, alias))
 	return q
 }
 
@@ -124,7 +134,7 @@ func (q Query) Max(fieldname string, aliases ...string) Query {
 	if len(aliases) > 0 {
 		alias = aliases[0]
 	}
-	q.fields = append(q.fields, fmt.Sprintf("max(\"%s\") AS %s", fieldname, alias))
+	q.fields = append(q.fields, fmt.Sprintf("max(\"%s\") AS \"%s\"", fieldname, alias))
 	return q
 }
 
@@ -134,7 +144,17 @@ func (q Query) Mean(fieldname string, aliases ...string) Query {
 	if len(aliases) > 0 {
 		alias = aliases[0]
 	}
-	q.fields = append(q.fields, fmt.Sprintf("mean(\"%s\") AS %s", fieldname, alias))
+	q.fields = append(q.fields, fmt.Sprintf("mean(\"%s\") AS \"%s\"", fieldname, alias))
+	return q
+}
+
+// Last returns a formated string of the field `last`.
+func (q Query) Last(fieldname string, aliases ...string) Query {
+	alias := fieldname
+	if len(aliases) > 0 {
+		alias = aliases[0]
+	}
+	q.fields = append(q.fields, fmt.Sprintf("last(\"%s\") AS \"%s\"", fieldname, alias))
 	return q
 }
 
@@ -144,7 +164,7 @@ func (q Query) CumulativeSum(function funcType, fieldname string, aliases ...str
 	if len(aliases) > 0 {
 		alias = aliases[0]
 	}
-	q.fields = append(q.fields, fmt.Sprintf("cumulative_sum(%s(\"%s\")) AS %s", function, fieldname, alias))
+	q.fields = append(q.fields, fmt.Sprintf("cumulative_sum(%s(\"%s\")) AS \"%s\"", function, fieldname, alias))
 	return q
 }
 
@@ -154,7 +174,7 @@ func (q Query) NonNegativeDerivative(function funcType, fieldname string, durati
 	if len(aliases) > 0 {
 		alias = aliases[0]
 	}
-	q.fields = append(q.fields, fmt.Sprintf("non_negative_derivative(%s(\"%s\"), %s) AS %s", function, fieldname, duration, alias))
+	q.fields = append(q.fields, fmt.Sprintf("non_negative_derivative(%s(\"%s\"), %s) AS \"%s\"", function, fieldname, duration, alias))
 	return q
 }