+++ influx/querybuilder_test.go
@@ -58,47 +58,93 @@ func TestQueryBuilder(t *testing.T) {
 			Query:    NewQuery().On("serie").Field("test", "mean"),
 			Expected: `SELECT mean("test") AS "test" FROM "serie"`,
 		}, {
+			Name:     "a basic query",
+			Query:    NewQuery().On("serie").Mean("f1", "f1").Limit(1),
+			Expected: `SELECT mean("f1") AS "f1" FROM "serie" LIMIT 1`,
+		}, {
 			Name:     "with a basic condition",
 			Query:    NewQuery().On("serie").Field("test", "mean").Where("condF", Equal, `"value"`),
 			Expected: `SELECT mean("test") AS "test" FROM "serie" WHERE condF = "value"`,
+		}, {
+			Name:     "with a basic condition",
+			Query:    NewQuery().On("serie").Mean("test", "test_alias").Where("condF", Equal, `"value"`),
+			Expected: `SELECT mean("test") AS "test_alias" FROM "serie" WHERE condF = "value"`,
 		}, {
 			Name:     "with a complex condition",
 			Query:    NewQuery().On("serie").Field("test", "mean").Where("cond1F", Equal, `"value"`).And("cond2F", MoreOrEqual, "time() - 3m"),
 			Expected: `SELECT mean("test") AS "test" FROM "serie" WHERE cond1F = "value" AND cond2F >= time() - 3m`,
 		}, {
+			Name:     "with a complex condition",
+			Query:    NewQuery().On("serie").Mean("test", "test_alias").Where("cond1F", Equal, `"value"`).And("cond2F", MoreOrEqual, "time() - 3m"),
+			Expected: `SELECT mean("test") AS "test_alias" FROM "serie" WHERE cond1F = "value" AND cond2F >= time() - 3m`,
+		}, {
 			Name:     "with an even more complex condition",
 			Query:    NewQuery().On("serie").Field("test", "mean").Where("cond1F", Equal, `"value"`).And("cond2F", MoreOrEqual, "time() - 3m").And("cond3F", Equal, "'app_id'"),
 			Expected: `SELECT mean("test") AS "test" FROM "serie" WHERE cond1F = "value" AND cond2F >= time() - 3m AND cond3F = 'app_id'`,
+		}, {
+			Name:     "with an even more complex condition",
+			Query:    NewQuery().On("serie").Mean("test", "test_alias").Where("cond1F", Equal, `"value"`).And("cond2F", MoreOrEqual, "time() - 3m").And("cond3F", Equal, "'app_id'"),
+			Expected: `SELECT mean("test") AS "test_alias" FROM "serie" WHERE cond1F = "value" AND cond2F >= time() - 3m AND cond3F = 'app_id'`,
 		}, {
 			Name:     "with a group by tag",
 			Query:    NewQuery().On("serie").Field("f1", "last").GroupByTag("tag1"),
 			Expected: `SELECT last("f1") AS "f1" FROM "serie" GROUP BY tag1`,
 		}, {
+			Name:     "with a group by tag",
+			Query:    NewQuery().On("serie").Last("f1", "f1_alias").GroupByTag("tag1"),
+			Expected: `SELECT last("f1") AS "f1_alias" FROM "serie" GROUP BY tag1`,
+		}, {
 			Name:     "with multiple group by time",
 			Query:    NewQuery().On("serie").Field("f1", "last").GroupByTag("tag1").GroupByTag("tag2"),
 			Expected: `SELECT last("f1") AS "f1" FROM "serie" GROUP BY tag1,tag2`,
+		}, {
+			Name:     "with multiple group by time",
+			Query:    NewQuery().On("serie").Last("f1", "f1_alias").GroupByTag("tag1").GroupByTag("tag2"),
+			Expected: `SELECT last("f1") AS "f1_alias" FROM "serie" GROUP BY tag1,tag2`,
 		}, {
 			Name:     "with a group by time",
 			Query:    NewQuery().On("serie").Field("f1", "last").GroupByTime(10 * time.Minute),
 			Expected: `SELECT last("f1") AS "f1" FROM "serie" GROUP BY time(10m0s)`,
 		}, {
+			Name:     "with a group by time",
+			Query:    NewQuery().On("serie").Last("f1", "f1_alias").GroupByTime(10 * time.Minute),
+			Expected: `SELECT last("f1") AS "f1_alias" FROM "serie" GROUP BY time(10m0s)`,
+		}, {
 			Name:     "with time, tags and fill",
 			Query:    NewQuery().On("serie").Field("f1", "last").GroupByTag("tag1").GroupByTag("tag2").GroupByTime(1 * time.Second).Fill(Previous),
 			Expected: `SELECT last("f1") AS "f1" FROM "serie" GROUP BY time(1s),tag1,tag2 fill(previous)`,
+		}, {
+			Name:     "with time, tags and fill",
+			Query:    NewQuery().On("serie").Last("f1", "f1_alias").GroupByTag("tag1").GroupByTag("tag2").GroupByTime(1 * time.Second).Fill(Previous),
+			Expected: `SELECT last("f1") AS "f1_alias" FROM "serie" GROUP BY time(1s),tag1,tag2 fill(previous)`,
 		}, {
 			Name:     "with an order by",
 			Query:    NewQuery().On("serie").Field("f1", "mean").OrderByTime("DESC"),
 			Expected: `SELECT mean("f1") AS "f1" FROM "serie" ORDER BY time DESC`,
 		}, {
+			Name:     "with an order by",
+			Query:    NewQuery().On("serie").Mean("f1", "f1_alias").OrderByTime("DESC"),
+			Expected: `SELECT mean("f1") AS "f1_alias" FROM "serie" ORDER BY time DESC`,
+		}, {
 			Name:     "with a limit",
 			Query:    NewQuery().On("serie").Field("f1", "mean").Limit(1),
 			Expected: `SELECT mean("f1") AS "f1" FROM "serie" LIMIT 1`,
+		}, {
+			Name:     "with a limit",
+			Query:    NewQuery().On("serie").Mean("f1", "f1_alias").Limit(1),
+			Expected: `SELECT mean("f1") AS "f1_alias" FROM "serie" LIMIT 1`,
 		}, {
 			Name: "a query with a subquery",
 			Query: NewQuery().OnSubquery(
 				NewQuery().On("serie").Field("biniou", "mean"),
 			).Field("biniou", "min"),
 			Expected: `SELECT min("biniou") AS "biniou" FROM (SELECT mean("biniou") AS "biniou" FROM "serie")`,
+		}, {
+			Name: "a query with a subquery",
+			Query: NewQuery().OnSubquery(
+				NewQuery().On("serie").Mean("biniou", "biniou_alias"),
+			).Min("biniou", "biniou_min_alias"),
+			Expected: `SELECT min("biniou") AS "biniou_min_alias" FROM (SELECT mean("biniou") AS "biniou_alias" FROM "serie")`,
 		},
 	}
 	for _, example := range examples {