+++ influx/CHANGELOG.md
@@ -5,6 +5,8 @@
 * Bump github.com/stretchr/testify from 1.6.1 to 1.7.0
 * Bump github.com/influxdata/influxdb from 1.8.3 to 1.9.5
 * Bump go version to 1.16
+* Deprecation of Field function
+* Adding specific field functions
 
 ## v1.0.0